requirements.txt

fastapi
uvicorn
beautifulsoup4
requests
redis
google-generativeai
pydantic
urllib3
python-dotenv
